requirements.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.10.0
bleach==5.0.1
capstone==5.0.0rc2
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
coloredlogs==15.0.1
cryptography==37.0.4
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
python-dotenv
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
Flask==2.2.1
flask-unsign==1.2.0
gitdb==4.0.9
GitHacker==1.1.1
GitPython==3.1.27
humanfriendly==10.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
plumbum==1.7.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pymongo==4.3.3
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.27.1
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.1
stack-data==0.3.0
termcolor==1.1.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.1
widgetsnbextension==3.6.1
openai
flask_jwt_extended
pandas
matplotlib
plotly
scipy
scikit-learn
functools
flask_mongoengine
